Camera System Rework
The camera has been entirely reworked to use Cinemachine Virtual Camera.
A camera manager object and script were added.

At the moment of the commit, the project requires a lot of additional setup.
Personal Recommendation:
In the main branch, on the owner's PC, rebuild the camera system and then upload the project entirely.