UI Map Rework
Added a secondary camera and sorted both cameras to be under a camera manager object.
Secondary camera feeds into a render texture object called World Camera. That RT is taken via a new script called "MinimapController" into the RAW Image in the canvas.

As of this commit, the secondary camera just shows the world in Macro in the corner, including the fog of war.

To Be Added:
Clone the world after it has been generated into another place and put the camera there, without the fog of war.
Work out how to mark territory with high belief.