Complete Rebuild of startup sequence
Game now starts with the menu, requests input parameters of size.
While menu is up, player cannot be moved.
As it stands, int.tryparse for some reason is unable to parse tmprougui.text Remains to be seen how to be dealt with.
The player also doesnt move to a new location upon re-generation of the world.